requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
Pillow
qrcode

